@@ -1,5 +1,5 @@
 {
-  "name": "livraria-requalificar-frontend",
+  "name": "emp-teste-frontend",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -18,10 +18,10 @@
     "install": "^0.13.0",
     "material-ui-popup-state": "^2.0.0",
     "npm": "^8.5.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0",
+    "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
